+++ src/chains/non-evm/bitcoin.js
@@ -1,0 +1,9 @@
+module.exports = {
+    SYMBOL: 'BTC',
+    NAME: 'Bitcoin',
+    DECIMALS: 8,
+    ASSET_API: 'https://api.coingecko.com/api/v3/coins/bitcoin',
+    FUNCTIONALITY_SUPPORT: {
+      getNativeAssetDetails: true,
+    },
+};