+++ src/chains/non-evm/tezos.js
@@ -1,0 +1,9 @@
+module.exports = {
+    SYMBOL: 'XTZ',
+    NAME: 'Tezos',
+    DECIMALS: 6,
+    ASSET_API: 'https://api.coingecko.com/api/v3/coins/tezos',
+    FUNCTIONALITY_SUPPORT: {
+      getNativeAssetDetails: true,
+    },
+};